Refactor so useDrag is just in the same folder not in kolibri-core (it was premature)
